requirements.txt
langchain
python_dotenv
anthropic
openai
pypdf
faiss-cpu
tiktoken